move cvelist2osv into subdirectory
