+++ vulnfeeds/conversion/cve5/common_test.go
@@ -1,4 +1,4 @@
-package cvelist2osv
+package cve5
 
 import (
 	"testing"
