+++ vulnfeeds/conversion/cve5/linux_extractor.go
@@ -1,4 +1,4 @@
-package cvelist2osv
+package cve5
 
 import (
 	"fmt"
