+++ vulnfeeds/conversion/cve5/testdata/TestExtractVersionInfo_CVE_with_lessThan.yaml
