+++ vulnfeeds/conversion/cve5/version_extraction_test.go
@@ -1,4 +1,4 @@
-package cvelist2osv
+package cve5
 
 import (
 	"reflect"
